// packages/executooor-ethers/src/types/Executor.ts
export type PlaceholderStruct = {
  to: AddressLike;
  data: BytesLike;
  offset: BigNumberish;
  length: BigNumberish;
  resOffset: BigNumberish;
};

export type PlaceholderStructOutput = [
  to: string,
  data: string,
  offset: bigint,
  length: bigint,
  resOffset: bigint
] & {
  to: string;
  data: string;
  offset: bigint;
  length: bigint;
  resOffset: bigint;
};

export interface ExecutorInterface extends Interface {
  getFunction(
    nameOrSignature:
      | "callWithPlaceholders4845164670"
      | "call_g0oyU7o"
      | "exec_606BaXt"
      | "transfer"
  ): FunctionFragment;

  encodeFunctionData(
    functionFragment: "callWithPlaceholders4845164670",
    values: [
      AddressLike,
      BigNumberish,
      BytesLike,
      BytesLike,
      PlaceholderStruct[]
    ]
  ): string;
  encodeFunctionData(
    functionFragment: "call_g0oyU7o",
    values: [AddressLike, BigNumberish, BytesLike, BytesLike]
  ): string;
  encodeFunctionData(
    functionFragment: "exec_606BaXt",
    values: [BytesLike[]]
  ): string;
  encodeFunctionData(
    functionFragment: "transfer",
    values: [AddressLike, BigNumberish]
  ): string;

  decodeFunctionResult(
    functionFragment: "callWithPlaceholders4845164670",
    data: BytesLike
  ): Result;
  decodeFunctionResult(
    functionFragment: "call_g0oyU7o",
    data: BytesLike
  ): Result;
  decodeFunctionResult(
    functionFragment: "exec_606BaXt",
    data: BytesLike
  ): Result;
  decodeFunctionResult(functionFragment: "transfer", data: BytesLike): Result;
}

export interface Executor extends BaseContract {
  connect(runner?: ContractRunner | null): Executor;
  waitForDeployment(): Promise<this>;

  interface: ExecutorInterface;

  queryFilter<TCEvent extends TypedContractEvent>(
    event: TCEvent,
    fromBlockOrBlockhash?: string | number | undefined,
    toBlock?: string | number | undefined
  ): Promise<Array<TypedEventLog<TCEvent>>>;
  queryFilter<TCEvent extends TypedContractEvent>(
    filter: TypedDeferredTopicFilter<TCEvent>,
    fromBlockOrBlockhash?: string | number | undefined,
    toBlock?: string | number | undefined
  ): Promise<Array<TypedEventLog<TCEvent>>>;

  on<TCEvent extends TypedContractEvent>(
    event: TCEvent,
    listener: TypedListener<TCEvent>
  ): Promise<this>;
  on<TCEvent extends TypedContractEvent>(
    filter: TypedDeferredTopicFilter<TCEvent>,
    listener: TypedListener<TCEvent>
  ): Promise<this>;

  once<TCEvent extends TypedContractEvent>(
    event: TCEvent,
    listener: TypedListener<TCEvent>
  ): Promise<this>;
  once<TCEvent extends TypedContractEvent>(
    filter: TypedDeferredTopicFilter<TCEvent>,
    listener: TypedListener<TCEvent>
  ): Promise<this>;

  listeners<TCEvent extends TypedContractEvent>(
    event: TCEvent
  ): Promise<Array<TypedListener<TCEvent>>>;
  listeners(eventName?: string): Promise<Array<Listener>>;
  removeAllListeners<TCEvent extends TypedContractEvent>(
    event?: TCEvent
  ): Promise<this>;

  callWithPlaceholders4845164670: TypedContractMethod<
    [
      target: AddressLike,
      value: BigNumberish,
      context: BytesLike,
      callData: BytesLike,
      placeholders: PlaceholderStruct[]
    ],
    [void],
    "payable"
  >;

  call_g0oyU7o: TypedContractMethod<
    [
      target: AddressLike,
      value: BigNumberish,
      context: BytesLike,
      callData: BytesLike
    ],
    [void],
    "payable"
  >;

  exec_606BaXt: TypedContractMethod<[data: BytesLike[]], [void], "payable">;

  transfer: TypedContractMethod<
    [recipient: AddressLike, amount: BigNumberish],
    [void],
    "payable"
  >;

  getFunction<T extends ContractMethod = ContractMethod>(
    key: string | FunctionFragment
  ): T;

  getFunction(
    nameOrSignature: "callWithPlaceholders4845164670"
  ): TypedContractMethod<
    [
      target: AddressLike,
      value: BigNumberish,
      context: BytesLike,
      callData: BytesLike,
      placeholders: PlaceholderStruct[]
    ],
    [void],
    "payable"
  >;
  getFunction(
    nameOrSignature: "call_g0oyU7o"
  ): TypedContractMethod<
    [
      target: AddressLike,
      value: BigNumberish,
      context: BytesLike,
      callData: BytesLike
    ],
    [void],
    "payable"
  >;
  getFunction(
    nameOrSignature: "exec_606BaXt"
  ): TypedContractMethod<[data: BytesLike[]], [void], "payable">;
  getFunction(
    nameOrSignature: "transfer"
  ): TypedContractMethod<
    [recipient: AddressLike, amount: BigNumberish],
    [void],
    "payable"
  >;

  filters: {};
}